Fixed an issue with Word 2013
When the opiton "Open e-mail attachments and other uneditbale files in reading view" is checked for Word 2013, any read-only document is opened in the "reading" view. In some circumenstances the Word 2013 will return the incorrect page number when "wdActiveEndPageNumber" is retrieved. This seems a Word 2013 specific issue. The workaround is to force the document to be opened in normal view by setting application.OpenAttachmentsInFullScreen = False. This setting is restored before the program terminates so no worries about messing up user settings.